pyproject.toml
[project]
name = "ami-orchestrator-root"
version = "0.0.0"
description = "Root orchestrator module scaffolding"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml==6.0.2",
    "resolvelib>=1.2.0",
    "packaging>=25.0",
    "google-auth-oauthlib==1.2.0",
]

[dependency-groups]
dev = [
    "pre-commit==4.3.0",
    "ruff==0.12.8",
    "mypy==1.17.1",
    "types-requests==2.32.0.20241016",
    "types-PyYAML==6.0.12.20250915",
    "requests==2.32.5",
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
    "pytest-timeout==2.4.0",
    "pytest-cov==6.2.1",
    "pytest-mock==3.15.1",
    "python-dotenv==1.1.1",
    "loguru==0.7.3",
    "pydantic==2.11.9",
    "pyyaml==6.0.2",
    "tiktoken==0.12.0",
    "uv==0.9.0",
    "zstandard==0.25.0",
    "google-auth==2.41.1",
    "google-api-python-client==2.185.0",
    "pillow==12.0.0",
    "cairosvg==2.7.1",
    "podman-compose==1.5.0",
    "nodeenv==1.9.1",
]

